auth-service/requirements.txt
Flask
psycopg2-binary
PyJWT
bcrypt
